python_env/pyproject.toml
[project]
name = "python-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "libclang>=18.1.1",
]
